Version 0.5, Bug fixes and improvements.
Fixed a bug that was allowing a lot of bosses to spawn in a boss fight.
Fixed a bug in the level editor preset created by the randomizer where
the value for the 1st boss was crashing the level editor.
Improvements in all random profiles.
Minor bug fixes.
Minor changes.